requirements.txt
Django==3.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
djangorestframework==3.11.0